Extract abstract base classes for code reuse
